marketing-bot/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rc-marketing-bot"
version = "0.1.0"
description = "AI-powered marketing bot for RC Investment Properties"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.40.0",
    "click>=8.1",
    "requests>=2.31",
    "beautifulsoup4>=4.12",
    "python-dotenv>=1.0",
    "rich>=13.0",
    "schedule>=1.2",
    "flask>=3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-mock>=3.12",
    "coverage>=7.0",
]

[project.scripts]
rc-bot = "src.cli:main"

[tool.setuptools.packages.find]
where = ["."]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[tool.coverage.run]
source = ["src"]
omit = [
    "tests/*",
    "src/web/app.py",
    "src/cli.py",
    "src/integrations/claude_client.py",
]

[tool.coverage.report]
fail_under = 85
show_missing = true
